python_bot/requirements.txt
aiogram==3.13.1
aiosqlite==0.19.0
python-dotenv==1.0.1
redis==5.1.0  # для будущего C++ обмена